Move tests to appropriate modules
