+++ src/parser.rs
@@ -299,3 +299,76 @@ impl<'s> Parser<'s> {
     }
 }
 
+#[cfg(test)]
+mod tests {
+    use error::*;
+    use super::Parser;
+
+    #[test]
+    fn test_unfold1() {
+        let mut p = Parser{input: "ab\r\n c", pos: 2};
+        assert_eq!(p.consume_char(), Some('c'));
+        assert_eq!(p.pos, 6);
+    }
+
+    #[test]
+    fn test_unfold2() {
+        let mut p = Parser{input: "ab\n\tc\nx", pos: 2};
+        assert_eq!(p.consume_char(), Some('c'));
+        assert_eq!(p.consume_char(), Some('\n'));
+        assert_eq!(p.consume_char(), Some('x'));
+    }
+
+    #[test]
+    fn test_consume_while() {
+        let mut p = Parser{input:"af\n oo:bar", pos: 1};
+        assert_eq!(p.consume_while(|x| x != ':'), "foo");
+        assert_eq!(p.consume_char(), Some(':'));
+        assert_eq!(p.consume_while(|x| x != '\n'), "bar");
+    }
+
+    #[test]
+    fn test_consume_while2() {
+        let mut p = Parser{input:"af\n oo\n\t:bar", pos: 1};
+        assert_eq!(p.consume_while(|x| x != ':'), "foo");
+        assert_eq!(p.consume_char(), Some(':'));
+        assert_eq!(p.consume_while(|x| x != '\n'), "bar");
+    }
+
+    #[test]
+    fn test_consume_while3() {
+        let mut p = Parser{input:"af\n oo:\n bar", pos: 1};
+        assert_eq!(p.consume_while(|x| x != ':'), "foo");
+        assert_eq!(p.consume_char(), Some(':'));
+        assert_eq!(p.consume_while(|x| x != '\n'), "bar");
+    }
+
+    #[test]
+    fn test_consume_only_char() {
+        let mut p = Parser{input:"\n \"bar", pos: 0};
+        assert!(p.consume_only_char('"'));
+        assert_eq!(p.pos, 3);
+        assert!(!p.consume_only_char('"'));
+        assert_eq!(p.pos, 3);
+        assert!(p.consume_only_char('b'));
+        assert_eq!(p.pos, 4);
+    }
+
+    #[test]
+    fn mismatched_begin_end_tags_returns_error() {
+        // Test for infinite loops as well
+        use std::sync::mpsc::{channel, RecvTimeoutError};
+        use std::time::Duration;
+        let mut p = Parser {input: "BEGIN:a\nBEGIN:b\nEND:a", pos: 0};
+
+        let (tx, rx) = channel();
+        ::std::thread::spawn(move|| { tx.send(p.consume_component()) });
+
+        match rx.recv_timeout(Duration::from_millis(50)) {
+            Err(RecvTimeoutError::Timeout) => assert!(false),
+            Ok(Err(VObjectError(VObjectErrorKind::ParserError{..}, _ ))) => assert!(true),
+            _ => assert!(false),
+        }
+    }
+
+}
